requirements.txt
# 屏幕翻译工具依赖包
PyQt6>=6.5.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
keyboard>=0.13.5
pyinstaller>=5.13.0
opencv-python>=4.8.1
numpy>=1.24.0
transformers>=4.21.0
torch>=2.0.0
sentencepiece>=0.1.99
psutil>=5.9.8
comtypes>=1.3.1
frida>=16.1.0
